Language option in CLI
In order to do that, tranlstion lookup has to be moved to output phase as jp/en
translation are usually not update to date. So if someone uses user_info from
cn/tw/kr and, there will be guns/equipments with empty strings as the name.
In the end the bogus result will be generated.